Database changes and Admin Add user UI codes
A few changes in database:
1. First name and last name for staff info was merged to one entity -
name.
2. New faculties added: msc_it and bba_management
3. New faculty added for staffs who are NOT counsellors: n/a. This can
be changed to be hidden from UI; but to be present in database instead
of 0 and one. we can program it something like, if faculty_id=n/a, do
this and do that.

all of these are open to change according to further discussions of
course.

Assumptions added:
The admin role is absolute and unchangeable, and single.
One counsellor can belong to one faculty only.
A counsellor is not divided between different degrees of the same
faculty such as bachelors in IT and masters in IT has the same
counsellor-faculty. The default counsellor faculty for any particular
subject would be bachelors. For example, a counsellor may have his/her
faculty listed as bachelors in IT, but in practice, is actually the
counsellor for both Bachelors AND Masters in IT.

Please find the database export file (lead_management.sql) under
web/WEB-INF/temp_database_changes